Move shared entities to practice/ package
Entities that appear in practice-layer protocol signatures are shared
vocabulary and must live in the shared library, not in any single
bounded context. This eliminates the TYPE_CHECKING upward dependency
from practice/repositories.py to bin/cli/entities.py.

Moved to practice/entities.py:
  Project, ProjectStatus, Confidence, ResearchTopic, Skillset

Remaining in bin/cli/entities.py (lifecycle-only, future consulting/ BC):
  DecisionEntry, EngagementEntry

practice/ now has zero imports from bin/cli/ at any level.

Closes the dependency inversion identified as prerequisite for #47.
