+++ tests/test_entity_store.py
@@ -9,7 +9,7 @@ from __future__ import annotations
 
 import pytest
 
-from bin.cli.entities import ProjectStatus
+from practice.entities import ProjectStatus
 from practice.store import EntityStore
 
 from .conftest import make_decision, make_project
